SquareGrid::Init

+++ QSweeper/Source/MineFactory.h
@@ -15,7 +15,7 @@ public:
 
 	template<class MineType>
 	typename std::enable_if<std::is_base_of<Mine, MineType>::value, std::unique_ptr<Mine>>::type
-	MakeMine(const glm::vec2& pos)
+	MakeMine(const glm::ivec2& pos)
 	{
 		//static_assert(0);
 		return std::make_unique<MineType>(pos, mineOptions_);
@@ -23,12 +23,12 @@ public:
 
 	std::unique_ptr<Mine> test1()
 	{
-		return MakeMine<SimpleMine>(glm::vec2(0), 1.2f);
+		return MakeMine<SimpleMine>(glm::ivec2(0), 1.2f);
 	}
 
 	std::unique_ptr<Mine> test2()
 	{
-		return MakeMine<SimpleMine>(glm::vec2(0));
+		return MakeMine<SimpleMine>(glm::ivec2(0));
 	}
 
 	void SetSimpleMineOptions(MineOptions::SimpleMineOptions opts)
